Pair programming - refactoring
